# Conflicts:
#   packages/db/src/seeders/20250721010102-seed-store.ts

+++ packages/db/src/seeders/20250721010102-seed-store.ts
@@ -1,10 +1,14 @@
 import { QueryInterface } from 'sequelize';
+import { v4 as uuidv4 } from 'uuid';
 import { StoreStatus } from '../types/enum';
 
 export default {
   async up(queryInterface: QueryInterface): Promise<void> {
+    const now = new Date();
+
     await queryInterface.bulkInsert('STORES', [
       {
+        uuid: uuidv4(),
         user_id: 2,
         store_name: 'Test Store 1',
         email: 'store1@example.com',
@@ -14,17 +18,18 @@ export default {
         logo_url: null,
         description: 'This is the first test store',
         status: StoreStatus.APPROVED,
-        created_at: new Date(),
-        updated_at: new Date(),
+        created_at: now,
+        updated_at: now,
         deleted_at: null,
       },
       {
+        uuid: uuidv4(),
         user_id: 3,
         store_name: 'Test Store 2',
         email: 'store2@example.com',
         phone: '0898765432',
         business_type: 'Wholesale',
-        website: null,
+        website: '-', // ไม่มีเว็บ ใช้ "-"
         logo_url: null,
         description: 'Second store linked to same bank account',
         status: StoreStatus.APPROVED,
@@ -36,6 +41,6 @@ export default {
   },
 
   async down(queryInterface: QueryInterface): Promise<void> {
-    await queryInterface.bulkDelete('STORES', {}, {});
+    await queryInterface.bulkDelete('STORES', { user_id: [2, 3] });
   },
 };
